1. The page "My Articles" has been created. 2. Edited validation of empty elements. 3. Removing articles has been implemented.

+++ src/main/java/com/gmail/eugeneknysh21/blog/services/impl/ArticleServiceImpl.java
@@ -99,8 +99,14 @@ public class ArticleServiceImpl implements ArticleService {
 
     @Override
     public boolean remove(Long id) {
-        articleRepository.deleteById(id);
-        return true;
+        Long userId = userService.getPrincipal().getId();
+        Article article = articleRepository.findById(id).orElseThrow(() ->
+                new NoSuchElementException("The article to delete was not found."));
+        if (article.getAuthor().getId().equals(userId)) {
+            articleRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
